pyproject.toml
# SPDX-FileCopyrightText: 2018-2025 Joonas Rautiola <mail@joinemm.dev>
# SPDX-License-Identifier: MPL-2.0
# https://git.joinemm.dev/miso-bot

[project]
name = "miso-bot"
version = "7.0"
description = "Multipurpose Discord bot"
readme = "README.md"
requires-python = ">=3.13"
license = "MPL-2.0"

dependencies = [
    "aiomysql>=0.2.0",
    "arrow>=1.3.0",
    "async-cse>=0.3.0",
    "beautifulsoup4>=4.13.5",
    "bleach>=6.2.0",
    "colorgram-py>=1.2.0",
    "discord-py[speed]>=2.6.3",
    "durations-nlp>=1.0.1",
    "humanize>=4.13.0",
    "jishaku>=2.6.0",
    "kdtree>=0.16",
    "loguru>=0.7.3",
    "lxml>=6.0.1",
    "markdownify>=1.2.0",
    "minestat>=2.6.3",
    "prometheus-async>=25.1.0",
    "python-dotenv>=1.1.1",
    "random-user-agent>=1.0.1",
    "redis>=6.4.0",
    "regex>=2025.9.1",
    "uvloop>=0.21.0",
]

[tool.basedpyright]
venvPath = ".devenv/state"
venv = "venv"
exclude = [
  "**/__pycache__"
]
executionEnvironments = [
  { root = "." }
]

reportMissingTypeArgument = false
reportMissingParameterType = false
reportUnknownParameterType = false
reportUnknownMemberType = false
reportUnannotatedClassAttribute = false
reportUnknownVariableType = false
reportExplicitAny = false
reportUnusedCallResult = false

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I001", "I002"]
ignore = ["E501"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
